Add files via upload
The primary objective of this project is to develop a  book recommendation system based on historical data to provide personalized book suggestions,aims to enhance the reading experiences of users by recommending books that align closely with their preferences and interests.

The project involves analyzing the book dataset obtained from Kaggle using Python and Jupyter Notebook. The analysis will utilize libraries such as pandas for data manipulation, cosine similarity is used  to find the similarity between the books,and the model will be saved using the pickle library. Streamlit will be used to create an interactive web application that allows users to input book details and recive the recommdations.